+++ packages/integration/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @vanilla-extract/integration
 
+## 8.0.1
+
+### Patch Changes
+
+- [#1545](https://github.com/vanilla-extract-css/vanilla-extract/pull/1545) [`965fd03`](https://github.com/vanilla-extract-css/vanilla-extract/commit/965fd03ff26dd324ec24734aa7700f1fe89bd483) Thanks [@askoufis](https://github.com/askoufis)! - Extend `esbuild` dependency range to include `0.25.x`
+
+## 8.0.0
+
+### Major Changes
+
+- [#1536](https://github.com/vanilla-extract-css/vanilla-extract/pull/1536) [`a8248be`](https://github.com/vanilla-extract-css/vanilla-extract/commit/a8248befac51aa51d771b9b22a46209b1fd1e3b3) Thanks [@askoufis](https://github.com/askoufis)! - Remove `createCompiler` function and `Compiler` and `CreateCompilerOptions` types
+
+  BREAKING CHANGE: These APIs have been moved to the new `@vanilla-extract/compiler` package
+
+### Minor Changes
+
+- [#1536](https://github.com/vanilla-extract-css/vanilla-extract/pull/1536) [`a8248be`](https://github.com/vanilla-extract-css/vanilla-extract/commit/a8248befac51aa51d771b9b22a46209b1fd1e3b3) Thanks [@askoufis](https://github.com/askoufis)! - Export `serializeVanillaModule` function
+
+### Patch Changes
+
+- [#1530](https://github.com/vanilla-extract-css/vanilla-extract/pull/1530) [`5f66abb`](https://github.com/vanilla-extract-css/vanilla-extract/commit/5f66abbd607e76d491bbb7b9bfe9c64c882a53e8) Thanks [@askoufis](https://github.com/askoufis)! - Extend `esbuild` dependency range to include `0.24.x`
+
+- [#1535](https://github.com/vanilla-extract-css/vanilla-extract/pull/1535) [`ec0b024`](https://github.com/vanilla-extract-css/vanilla-extract/commit/ec0b024fd19c133c233445f9e860626d104f9d97) Thanks [@askoufis](https://github.com/askoufis)! - Fixes an error that sometimes occurred when loading a page via the astro dev server
+
+- Updated dependencies [[`c432ff3`](https://github.com/vanilla-extract-css/vanilla-extract/commit/c432ff33a8aead2c94fa6a4fcc9fcb1d7990427d)]:
+  - @vanilla-extract/css@1.17.1
+
+## 7.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`fd673f6`](https://github.com/vanilla-extract-css/vanilla-extract/commit/fd673f658da3b0019d19a0c7d6d9208d1685ff5b), [`fd673f6`](https://github.com/vanilla-extract-css/vanilla-extract/commit/fd673f658da3b0019d19a0c7d6d9208d1685ff5b), [`4abfc0b`](https://github.com/vanilla-extract-css/vanilla-extract/commit/4abfc0b47c675f9a7bfdb4c3316b1c62d2b8e558), [`fd673f6`](https://github.com/vanilla-extract-css/vanilla-extract/commit/fd673f658da3b0019d19a0c7d6d9208d1685ff5b)]:
+  - @vanilla-extract/babel-plugin-debug-ids@1.2.0
+  - @vanilla-extract/css@1.17.0
+
 ## 7.1.11
 
 ### Patch Changes