+++ packages/next-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @vanilla-extract/next-plugin
 
+## 2.4.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vanilla-extract/webpack-plugin@2.3.18
+
+## 2.4.9
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vanilla-extract/webpack-plugin@2.3.17
+
+## 2.4.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vanilla-extract/webpack-plugin@2.3.16
+
 ## 2.4.7
 
 ### Patch Changes